D2-12_Django_News_Project/requirements.txt
Django==3.1.7
django-allauth==0.44.0
django-apscheduler==0.5.2
django-filter==2.4.0
